requirements.txt
wheel
pandas
notebook
matplotlib
python-binance


